requirements.txt
streamlit==1.28.1
langchain==0.0.340
langchain-community==0.0.1
langchain-openai==0.0.2
chromadb==0.4.18
pymupdf==1.23.8
pdfplumber==0.10.3
openai==1.3.7
python-dotenv==1.0.0
sentence-transformers==2.2.2
tiktoken==0.5.2